Implement a function that allow us to add a compare link to any specific places.
Note, that because YITH WooCommerce Compare Premium plugin doesn't provide
a way to 'add' a comapre link manually, so it is always be hooked under
'woocommerce_after_shop_loop_item' action.

In order to custom it, we have to duplicate its code here so we can modify,
and execute this function in a specific place.